Fix logfiles.oper flag not working
When instantiating the Logfiles object, we need to distinguish two
cases:

When --logfiles option is specified on the command-line, we must load
all logfiles that are in the database - since the user needs to see all
existing logfiles, even those that are not operational.

When --logfiles option is not specified, then we only want to load
operational logfiles for processing.  Therefore the 'db' parameter to
Logfiles constructor needs to be filtered for this condition.